siim_yuji/requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
albumentations==0.4.6
antlr4-python3-runtime==4.8
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
configparser==5.0.2
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
docker-pycreds==0.4.0
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.6.1
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gspread==3.7.0
h5py==3.1.0
httplib2==0.19.1
huggingface-hub==0.0.8
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.5.0
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.4.2
multidict==5.1.0
munch==2.5.0
networkx==2.5.1
numba==0.53.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
omegaconf==2.1.0
opencv-python==4.5.2.52
opencv-python-headless==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pathtools==0.1.2
Pillow==8.2.0
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.17.2
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyDeprecate==0.3.0
pydicom==2.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.2.8
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
sentry-sdk==1.1.0
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
slackweb==1.0.5
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2021.4.8
timm==0.4.9
tokenizers==0.10.3
torch==1.8.1+cu111
torchaudio==0.8.1
torchmetrics==0.4.1
torchvision==0.9.1+cu111
tqdm==4.61.0
transformers==4.6.1
typing-extensions==3.7.4.3
urllib3==1.26.5
wandb==0.10.31
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
